+++ hack/.packages
@@ -41,6 +41,7 @@ k8s.io/kops/pkg/dns
 k8s.io/kops/pkg/edit
 k8s.io/kops/pkg/featureflag
 k8s.io/kops/pkg/flagbuilder
+k8s.io/kops/pkg/instancegroups
 k8s.io/kops/pkg/kubeconfig
 k8s.io/kops/pkg/model
 k8s.io/kops/pkg/model/awsmodel
